modified code

+++ admin/includes/view_all_posts.php
@@ -1,84 +1,48 @@
 <?php
 
-include ("delete_modal.php");
+include("delete_modal.php");
 
+if (isset($_POST['checkBoxArray']) && isset($_POST['bulk_options'])) {
 
+    $bulk_options = $_POST['bulk_options'];
 
-if (isset($_POST['checkBoxArray'])) {
+    // Ensure the user selected an action
+    if ($bulk_options !== "") {
+        foreach ($_POST['checkBoxArray'] as $postValueId) {
 
-    foreach ($_POST['checkBoxArray'] as $postValueId) {
+            switch ($bulk_options) {
+                case 'published':
+                    $query = "UPDATE posts SET post_status = '{$bulk_options}' WHERE post_id = {$postValueId} ";
+                    $update_to_published_status = mysqli_query($connection, $query);
+                    confirmQuery($update_to_published_status);
+                    break;  
 
-        $bulk_options = $_POST['bulk_options'];
-        switch ($bulk_options) {
-            case 'published':
-                $query = "UPDATE posts SET post_status = '{$bulk_options}' WHERE post_id ={$postValueId} ";
-                $update_to_published_status = mysqli_query($connection, $query);
-                confirmQuery($update_to_published_status);
-                break;
+                case 'draft':
+                    $query = "UPDATE posts SET post_status = '{$bulk_options}' WHERE post_id = {$postValueId} ";
+                    $update_to_draft_status = mysqli_query($connection, $query);
+                    confirmQuery($update_to_draft_status);
+                    break;  
 
-            case 'draft':
-                $query = "UPDATE posts SET post_status = '{$bulk_options}' WHERE post_id ={$postValueId} ";
-                $update_to_draft_status = mysqli_query($connection, $query);
-                confirmQuery($update_to_draft_status);
-                break;
+                case 'delete':
+                    // Delete the post and its associated comments
+                    $query = "DELETE FROM posts WHERE post_id = {$postValueId} ";
+                    $update_to_delete_status = mysqli_query($connection, $query);
+                    confirmQuery($update_to_delete_status);
 
-            case 'delete':
-                $query = "DELETE FROM posts WHERE post_id ={$postValueId} ";
-                $update_to_delete_status = mysqli_query($connection, $query);
-                confirmQuery($update_to_delete_status);
-                break;
-
-            case 'clone':
-                $query = "SELECT * FROM posts WHERE post_id = '{$postValueId}' ";
-                $select_post_query = mysqli_query($connection, $query);
-
-                while ($row = mysqli_fetch_array($select_post_query)) {
-                    $post_title = $row['post_title'];
-                    $post_category_id = $row['post_category_id'];
-                    $post_date = $row['post_date'];
-                    $post_author = $row['post_author'];
-                    $post_user = $row['post_user'];
-                    $post_status = $row['post_status'];
-                    $post_image = $row['post_image'];
-                    $post_tags = $row['post_tags'];
-                    $post_content = $row['post_content'];
-
-                    if (!empty($post_tags)) {
-                        $post_tags = "No tags";
-                    }
-                }
-
-                // INSERT INTO query
-                $query = "INSERT INTO posts(post_category_id, post_title, post_author,post_user, post_date, post_image, post_content, post_tags, post_status) ";
-                $query .= "VALUES({$post_category_id}, '{$post_title}', '{$post_author}', '{$post_user}', now(), '{$post_image}', '{$post_content}', '{$post_tags}', '{$post_status}')";
-
-                // Execute the query
-                $copy_query = mysqli_query($connection, $query);
-
-                // Check if query executed successfully
-                if (!$copy_query) {
-                    die("QUERY FAILED" . mysqli_error($connection));
-                }
+                    // DELETE associated comments as well
+                    $query = "DELETE FROM comments WHERE comment_post_id = {$postValueId} ";
+                    $delete_comments_query = mysqli_query($connection, $query);
+                    confirmQuery($delete_comments_query);
+                    break;  
+            }
         }
-        break;
-
-
-
-
-
-
-
-
-
-
     }
 }
 
 ?>
-<form action="" method='post'>
 
+<form action="" method="post">
     <table class="table table-bordered table-hover">
-
         <div id="bulkOptionsContainer" class="col-xs-4">
             <select class="form-control" name="bulk_options" id="">
                 <option value="">Select Option</option>
@@ -108,12 +72,13 @@ if (isset($_POST['checkBoxArray'])) {
                 <th>View Post</th>
                 <th>Edit</th>
                 <th>Delete</th>
+                <th>Views</th>
             </tr>
         </thead>
 
         <tbody>
             <?php
-            $query = "SELECT * FROM posts ";
+            $query = "SELECT * FROM posts";
             $select_posts = mysqli_query($connection, $query);
             while ($row = mysqli_fetch_assoc($select_posts)) {
                 $post_id = $row['post_id'];
@@ -130,16 +95,10 @@ if (isset($_POST['checkBoxArray'])) {
 
                 echo "<tr>"; ?>
 
-                <td><input class='checkBoxes' type='checkbox' name='checkBoxArray[]' value='<?php echo $post_id; ?>'></td>
-                <td>
-                    <?php echo $post_id; ?>
-                </td>
-                <td>
-                    <?php echo isset($post_user) ? $post_user : $post_author; ?>
-                </td>
-                <td>
-                    <?php echo $post_title; ?>
-                </td>
+                <td><input class="checkBoxes" type="checkbox" name="checkBoxArray[]" value="<?php echo $post_id; ?>"></td>
+                <td><?php echo $post_id; ?></td>
+                <td><?php echo isset($post_user) ? $post_user : $post_author; ?></td>
+                <td><?php echo $post_title; ?></td>
 
                 <?php
                 $query = "SELECT * FROM categories WHERE cat_id = {$post_category_id}";
@@ -150,13 +109,9 @@ if (isset($_POST['checkBoxArray'])) {
                     echo "<td>{$cat_title}</td>";
                 }
                 ?>
-                <td>
-                    <?php echo $post_status; ?>
-                </td>
+                <td><?php echo $post_status; ?></td>
                 <td><img width="100" src="../images/<?php echo $post_image; ?>" alt="image"></td>
-                <td>
-                    <?php echo $post_tags; ?>
-                </td>
+                <td><?php echo $post_tags; ?></td>
 
                 <?php
                 $query = "SELECT * FROM comments WHERE comment_post_id = $post_id";
@@ -166,7 +121,6 @@ if (isset($_POST['checkBoxArray'])) {
                 echo "<td><a href='post_comments.php?id={$post_id}'>{$comment_count}</a></td>";
 
                 echo "<td>" . $post_date . "</td>";
-
                 echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
                 echo " <td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
                 echo "<td><a rel='{$post_id}' href='javascript:void(0)' class='delete_link'>Delete</a></td>";
@@ -182,31 +136,35 @@ if (isset($_POST['checkBoxArray'])) {
 if (isset($_GET['delete'])) {
     $the_post_id = $_GET['delete'];
 
-    $query = "DELETE FROM posts WHERE post_id = {$the_post_id} ";
-    $delete_query = mysqli_query($connection, $query);
-    header("Location:posts.php");
+    // Delete comments associated with the post
+    $query = "DELETE FROM comments WHERE comment_post_id = {$the_post_id}";
+    $delete_comments_query = mysqli_query($connection, $query);
 
+    // Delete the post
+    $query = "DELETE FROM posts WHERE post_id = {$the_post_id}";
+    $delete_query = mysqli_query($connection, $query);
+
+    header("Location:posts.php");
 }
+
 if (isset($_GET['reset'])) {
     $the_post_id = $_GET['reset'];
 
     $query = "UPDATE posts SET post_views_count= 0 WHERE post_id =" . mysqli_real_escape_string($connection, $_GET['reset']) . " ";
     $reset_query = mysqli_query($connection, $query);
     header("Location:posts.php");
-
 }
 ?>
+
 <script>
     $(document).ready(function () {
         $(".delete_link").on('click', function () {
             var id = $(this).attr("rel");
-            var delete_url = "posts.php?delete=" + id; // Corrected concatenation
+            var delete_url = "posts.php?delete=" + id; 
 
             $(".modal_delete_link").attr("href", delete_url);
 
             $("#myModal").modal('show');
         });
     });
-
-
 </script>
